requirements.txt
Pillow
numpy
torch
torchvision
libsvm-official
matplotlib
prettytable